+++ test-crates/with-data/with_data.data/extra-src/sneaky_extra_module.py
@@ -1,0 +1,2 @@
+def silent_paws() -> str:
+    return "🐈"